Add new-chat features Complete message send and user select js
