+++ src/lib/messages.svelte
@@ -1,228 +1,161 @@
 <script lang="ts">
-    import { onMount, onDestroy } from "svelte";
-
+	import Messages from './messages.svelte';
+    import Sendbox from "./sendbox.svelte";
+    import { onMount, onDestroy, afterUpdate } from "svelte";
     let newMessage = "";
-    let targetUser = {
-        userId: 1,
+    export let targetUser = {
+        id: 1,
         name: "Allen Panda",
         avartar: "https://mdbcdn.b-cdn.net/img/new/avatars/1.webp",
         status: "Online",
+        messageGroup: [],
     };
-    let messages = [
-        {
-            id: 1,
-            time: "19-08-23 7:32 pm",
-            messages: [
-                {
-                    id: 111312,
-                    user: {
-                        userId: 0,
-                        name: "Me",
-                        avartar:
-                            "https://mdbcdn.b-cdn.net/img/new/avatars/2.webp",
-                    },
-                    message: "Test message from me",
-                },
-                {
-                    id: 111232,
-                    user: {
-                        userId: 1,
-                        name: "Allen Panda",
-                        avartar:
-                            "https://mdbcdn.b-cdn.net/img/new/avatars/1.webp",
-                    },
-                    message: "Good evening",
-                },
-            ],
-        },
-        {
-            id: 2,
-            time: "Monday 12:32 pm",
-            messages: [
-                {
-                    id: 114152,
-                    user: {
-                        userId: 0,
-                        name: "Me",
-                        avartar:
-                            "https://mdbcdn.b-cdn.net/img/new/avatars/2.webp",
-                    },
-                    message: "How is it going?",
-                },
-                {
-                    id: 156654,
-                    user: {
-                        userId: 1,
-                        name: "Allen Panda",
-                        avartar:
-                            "https://mdbcdn.b-cdn.net/img/new/avatars/1.webp",
-                    },
-                    message: "Great!",
-                },
-            ],
-        },
-        {
-            id: 3,
-            time: "Yesterday 15:11 pm",
-            messages: [
-                {
-                    id: 125676,
-                    user: {
-                        userId: 1,
-                        name: "Allen Panda",
-                        avartar:
-                            "https://mdbcdn.b-cdn.net/img/new/avatars/1.webp",
-                    },
-                    message: "See u tmr",
-                },
-                {
-                    id: 215315,
-                    user: {
-                        userId: 0,
-                        name: "Me",
-                        avartar:
-                            "https://mdbcdn.b-cdn.net/img/new/avatars/2.webp",
-                    },
-                    message: "See u",
-                },
-            ],
-        },
-        {
-            id: 4,
-            time: "09:00 pm",
-            messages: [
-                {
-                    id: 125326,
-                    user: {
-                        userId: 1,
-                        name: "Allen Panda",
-                        avartar:
-                            "https://mdbcdn.b-cdn.net/img/new/avatars/1.webp",
-                    },
-                    message: "where are u?",
-                },
-                {
-                    id: 154314,
-                    user: {
-                        userId: 1,
-                        name: "Allen Panda",
-                        avartar:
-                            "https://mdbcdn.b-cdn.net/img/new/avatars/1.webp",
-                    },
-                    message: "I just arrived",
-                },
-                {
-                    id: 765365,
-                    user: {
-                        userId: 0,
-                        name: "Me",
-                        avartar:
-                            "https://mdbcdn.b-cdn.net/img/new/avatars/2.webp",
-                    },
-                    message: "On the bus, coming",
-                },
-            ],
-        },
-    ];
 
-    onMount(async () => {});
+    onMount(async () => {
+        console.log('test');
+        scrollToBottom();
+    });
 
     // Unsubscribe from realtime messages
     onDestroy(() => {});
 
-    function sendMessage() {
-        //   const data = {
-        //     text: newMessage,
-        //     user: $currentUser.id,
-        //   };
-        //   const createdMessage = await pb.collection('messages').create(data);
-        //   newMessage = '';
-    }
+    afterUpdate(() => {
+        scrollToBottom();
+    });
+
+    const scrollToBottom = () => {
+        var domWrapper = document.querySelector(".modal-body");
+        domWrapper.scrollTo(0, 9999999);
+        console.log(domWrapper.scrollTop);
+    };
+
+    const sendMessage = (e) => {
+        let content = e.detail;
+        console.log("send msg", content);
+
+        if(targetUser.messageGroup && targetUser.messageGroup.length > 0){
+            var lastMsgGroupItem = targetUser.messageGroup[targetUser.messageGroup.length - 1]
+            if(lastMsgGroupItem.messages.length > 0){
+                var newMsg = {
+                            id: Math.random(),
+                            user: {
+                                userId: 0,
+                                name: "Me",
+                                avartar:
+                                    "https://mdbcdn.b-cdn.net/img/new/avatars/2.webp",
+                            },
+                            message: content,
+                        };
+                targetUser.messageGroup[targetUser.messageGroup.length - 1].messages = [...targetUser.messageGroup[targetUser.messageGroup.length - 1].messages, newMsg ]
+            }
+            else{
+                var newMsgGroupItem = {
+                    id: Math.random(),
+                    time: "Just now",
+                    messages: [
+                        {
+                            id: Math.random(),
+                            user: {
+                                userId: 0,
+                                name: "Me",
+                                avartar:
+                                    "https://mdbcdn.b-cdn.net/img/new/avatars/2.webp",
+                            },
+                            message: content,
+                        },
+                    ],
+                };
+                
+            targetUser.messageGroup = [...targetUser.messageGroup, newMsgGroupItem];
+            }
+            
+        }
+        else{
+            var newMsgGroup = [
+                {
+                    id: Math.random(),
+                    time: "Just now",
+                    messages: [
+                        {
+                            id: Math.random(),
+                            user: {
+                                userId: 0,
+                                name: "Me",
+                                avartar:
+                                    "https://mdbcdn.b-cdn.net/img/new/avatars/2.webp",
+                            },
+                            message: content,
+                        },
+                    ],
+                }
+            ]
+            targetUser.messageGroup = newMsgGroup;
+        }
+    };
 </script>
 
-<div class="chatbox">
-    <div class="modal-dialog-scrollable">
-        <div class="modal-content">
-            <div class="msg-head">
-                <div class="row">
-                    <div class="col-12">
-                        <div class="d-flex align-items-center">
-                            <div class="flex-shrink-0">
-                                <img
-                                    class="rounded-circle"
-                                    src={targetUser.avartar}
-                                    alt="avatar"
-                                    width="60px"
-                                />
-                            </div>
-                            <div class="flex-grow-1 ms-3">
-                                <h3>{targetUser.name}</h3>
-                                <p>{targetUser.status}</p>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-
-            <div class="modal-body">
-                <div class="msg-body">
-                    {#each messages as messageGroup (messageGroup.id)}
-                        <span class="time">{messageGroup.time}</span>
-                        {#each messageGroup.messages as message (message.id)}
-                            <ul>
-                                <li
-                                    class={message.user.userId ===
-                                    targetUser.userId
-                                        ? "sender"
-                                        : "reply"}
-                                >
-                                    <img
-                                        class="rounded-circle"
-                                        src={message.user?.avartar}
-                                        alt="avatar"
-                                        width="50px"
-                                    />
-                                    <div class="msg-content">
-                                        {#if message.user.userId ===
-                                            targetUser.userId}
-                                            <small>
-                                                {message.user?.name}
-                                            </small>
-                                        {/if}
-                                        
-                                        <p class="msg-text">
-                                            {message.message}
-                                        </p>
-                                    </div>
-                                </li>
-                            </ul>
-                        {/each}
-                    {/each}
-                </div>
-            </div>
-
-            <div class="send-box">
-                <form action="">
-                    <input
-                        type="text"
-                        class="form-control"
-                        aria-label="message…"
-                        placeholder="Write message…"
-                    />
-
-                    <button type="button"
-                        ><i class="fa fa-paper-plane" aria-hidden="true" /> Send</button
-                    >
-                </form>
-            </div>
+<div class="msg-head">
+    <div class="d-flex align-items-center">
+        <div class="flex-shrink-0">
+            <img
+                class="rounded-circle"
+                src={targetUser.avartar}
+                alt="avatar"
+                width="60px"
+            />
+        </div>
+        <div class="flex-grow-1 ms-3">
+            <h3>{targetUser.name}</h3>
+            <p>{targetUser.status}</p>
         </div>
     </div>
 </div>
 
+<div class="modal-dialog-scrollable">
+    <div class="modal-content">
+        <div class="modal-body">
+            <div class="msg-body">
+                {#if targetUser && targetUser.messageGroup}
+                {#each targetUser.messageGroup as messageGroup (messageGroup.id)}
+                    <span class="time">{messageGroup.time}</span>
+                    {#each messageGroup.messages as message (message.id)}
+                        <ul>
+                            <li
+                                class={message.user.userId === targetUser.id
+                                    ? "sender"
+                                    : "reply"}
+                            >
+                                <img
+                                    class="rounded-circle"
+                                    src={message.user?.avartar}
+                                    alt="avatar"
+                                    width="50px"
+                                />
+                                <div class="msg-content">
+                                    {#if message.user.userId === targetUser.id}
+                                        <small>
+                                            {message.user?.name}
+                                        </small>
+                                    {/if}
+
+                                    <p class="msg-text">
+                                        {message.message}
+                                    </p>
+                                </div>
+                            </li>
+                        </ul>
+                    {/each}
+                {/each}
+                {/if}
+            </div>
+        </div>
+    </div>
+</div>
+<Sendbox on:messageSent={sendMessage} />
+
 <style>
     ul,
-    li
-    {
+    li {
         margin: 0;
         padding: 0;
         border: 0;
@@ -249,11 +182,11 @@
         margin-bottom: 0;
     }
 
-    .chatbox {
+    .modal-dialog-scrollable {
         width: auto;
-        overflow: hidden;
-        height: 100%;
-        border-left: 1px solid #ccc;
+        overflow-y: auto;
+        height: 38rem;
+        /* border-left: 1px solid #ccc; */
     }
 
     .msg-head h3 {
@@ -294,18 +227,18 @@
     }
 
     .msg-body ul li.sender img {
-        display:inline-block;
+        display: inline-block;
         position: relative;
     }
 
     .msg-body ul li.sender div.msg-content {
-        display:inline-block;
+        display: inline-block;
         position: relative;
         margin-left: 5px;
     }
 
     .msg-body ul li.sender small {
-        display:flex;
+        display: flex;
         width: 100%;
         position: relative;
     }
@@ -333,7 +266,7 @@
         line-height: 1.5;
         font-weight: 400;
         padding: 15px;
-        background: #f5f5f5;
+        background: #e8e4e4;
         display: inline-block;
         border-bottom-left-radius: 10px;
         border-top-right-radius: 10px;
@@ -364,7 +297,7 @@
         display: inline-block;
     }
 
-    .msg-body ul li.reply:before {        
+    .msg-body ul li.reply:before {
         float: right;
         display: block;
         clear: both;
@@ -419,91 +352,8 @@
         padding: 4px;
     }
 
-    .send-box {
-        padding: 15px;
-        border-top: 1px solid #ccc;
-    }
-
-    .send-box form {
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-        margin-bottom: 15px;
-    }
-
-    .send-box .form-control {
-        display: block;
-        width: 85%;
-        padding: 0.375rem 0.75rem;
-        font-size: 14px;
-        font-weight: 400;
-        line-height: 1.5;
-        color: #222;
-        background-color: #fff;
-        background-clip: padding-box;
-        border: 1px solid #ccc;
-        -webkit-appearance: none;
-        -moz-appearance: none;
-        appearance: none;
-        border-radius: 0.25rem;
-        transition: border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
-    }
-
-    .send-box button {
-        border: none;
-        background: #3867d6;
-        padding: 0.375rem 5px;
-        color: #fff;
-        border-radius: 0.25rem;
-        font-size: 14px;
-        font-weight: 400;
-        width: 24%;
-        margin-left: 1%;
-    }
-
-    .send-box button i {
-        margin-right: 5px;
-    }
-
-    .send-btns .button-wrapper {
-        position: relative;
-        width: 125px;
-        height: auto;
-        text-align: left;
-        margin: 0 auto;
-        display: block;
-        background: #f6f7fa;
-        border-radius: 3px;
-        padding: 5px 15px;
-        float: left;
-        margin-right: 5px;
-        margin-bottom: 5px;
-        overflow: hidden;
-    }
-
-    .send-btns .button-wrapper span.label {
-        position: relative;
-        z-index: 1;
-        display: -webkit-box;
-        display: -ms-flexbox;
-        display: flex;
-        -webkit-box-align: center;
-        -ms-flex-align: center;
-        align-items: center;
-        width: 100%;
-        cursor: pointer;
-        color: #343945;
-        font-weight: 400;
-        text-transform: capitalize;
-        font-size: 13px;
-    }
-
-    button:focus {
-        outline: 0;
-    }
-
     @media (max-width: 767px) {
-        .chatbox {
+        .modal-dialog-scrollable {
             width: 100%;
             position: absolute;
             left: 1000px;
@@ -521,12 +371,7 @@
         .msg-head .flex-shrink-0 img {
             height: 30px;
         }
-        .send-box button {
-            width: 28%;
-        }
-        .send-box .form-control {
-            width: 70%;
-        }
+
         .msg-body ul li.sender p {
             font-size: 13px;
             padding: 8px;
