+++ src/lib/sidebar.svelte
@@ -1,5 +1,5 @@
 <script>
-  import { base } from '$app/paths'; // for github pages
+  import { base } from "$app/paths"; // for github pages
   import Bell from "svelte-bootstrap-icons/lib/Bell.svelte";
   import ChatDots from "svelte-bootstrap-icons/lib/ChatDots.svelte";
 
@@ -11,16 +11,30 @@
 </script>
 
 <div class="d-flex flex-column flex-shrink-0 bg-light" style="width: 4.5rem;">
-    <ul class="nav nav-pills nav-flush flex-column mb-auto text-center">
-      <li>
-        <a href="{base}/" class="nav-link py-3 border-bottom" title="" data-bs-toggle="tooltip" data-bs-placement="right" data-bs-original-title="Chat">
-          <ChatDots {size} {width} {height} {color}/>
-        </a>
-      </li>
-      <li>
-        <a href="{base}/quiz" class="nav-link py-3 border-bottom" title="" data-bs-toggle="tooltip" data-bs-placement="right" data-bs-original-title="quiz">
-          <Bell {size} {width} {height} {color}/>
-        </a>
-      </li>
-    </ul>
-  </div>
+  <ul class="nav nav-pills nav-flush flex-column mb-auto text-center">
+    <li>
+      <a
+        href="{base}/"
+        class="nav-link py-3 border-bottom"
+        title=""
+        data-bs-toggle="tooltip"
+        data-bs-placement="right"
+        data-bs-original-title="Chat"
+      >
+        <ChatDots {size} {width} {height} {color} />
+      </a>
+    </li>
+    <li>
+      <a
+        href="{base}/quiz"
+        class="nav-link py-3 border-bottom"
+        title=""
+        data-bs-toggle="tooltip"
+        data-bs-placement="right"
+        data-bs-original-title="quiz"
+      >
+        <Bell {size} {width} {height} {color} />
+      </a>
+    </li>
+  </ul>
+</div>
